+++ src/pages/Voucher/add/index.tsx
@@ -2,13 +2,19 @@ import { BreadCrumb } from "@src/components/atoms/Breadcrumb";
 import Content from "@src/components/molecules/content";
 import { CreateVoucherForm } from "@src/components/templates/admin/voucher/voucher-add";
 import { useCreateVoucher, useUpdateVoucher } from "@src/service/voucher";
+import { useLocation } from "react-router-dom";
 
 const VoucherAdd = () => {
   const { mutateAsync: mutate, isLoading } = useCreateVoucher();
   const { mutateAsync: update, isLoading: isUpdating } = useUpdateVoucher();
+  const location = useLocation();
+  const { state } = location;
   return (
     <>
-      <BreadCrumb name="Voucher & Promocode / Add" />
+      <BreadCrumb
+        name="Voucher & Promocode"
+        subname={state == null ? "Add New" : "Update"}
+      />
       <Content>
         <CreateVoucherForm
           mutate={mutate}